+++ Api/AntScape/src/main/java/es/iespto/algyjmcg/AntScape/infrastructure/adapter/primary/v2/NestLevelV2Controller.java
@@ -33,7 +33,7 @@ public class NestLevelV2Controller {
 		
 		BigDecimal res = findById.getMultiplier().multiply(BigDecimal.valueOf(findById.getProduction()));
 		
-		findById.setProduction(res.intValue());
+		findById.setProduction(res.doubleValue());
 		
 		boolean update = mainService.update(findById);
 		
@@ -90,7 +90,7 @@ class NestLevelSaveInputDTO {
 	private Integer level;
 	private BigDecimal multiplier;
 	private String name;
-	private Integer production;
+	private Double production;
 	private Integer id_nest;
 	
 	public NestLevelSaveInputDTO() {
@@ -127,10 +127,10 @@ class NestLevelSaveInputDTO {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Integer getProduction() {
+	public Double getProduction() {
 		return production;
 	}
-	public void setProduction(Integer production) {
+	public void setProduction(Double production) {
 		this.production = production;
 	}
 	public Integer getId_nest() {
